+++ GRDB/Core/DatabaseSnapshot.swift
@@ -126,20 +126,16 @@ extension DatabaseSnapshot {
         case .mainQueue:
             if DispatchQueue.isMain {
                 do {
-                    if let value = try read(observation.fetchFirst) {
-                        onChange(value)
-                    }
+                    try onChange(serializedDatabase.reentrantSync(observation.fetchValue))
                 } catch {
                     onError(error)
                 }
             } else {
                 serializedDatabase.async { db in
-                    let result = Result { try observation.fetchFirst(db) }
+                    let result = Result { try observation.fetchValue(db) }
                     DispatchQueue.main.async {
                         do {
-                            if let value = try result.get() {
-                                onChange(value)
-                            }
+                            try onChange(result.get())
                         } catch {
                             onError(error)
                         }
@@ -148,12 +144,10 @@ extension DatabaseSnapshot {
             }
         case let .async(onQueue: queue):
             serializedDatabase.async { db in
-                let result = Result { try observation.fetchFirst(db) }
+                let result = Result { try observation.fetchValue(db) }
                 queue.async {
                     do {
-                        if let value = try result.get() {
-                            onChange(value)
-                        }
+                        try onChange(result.get())
                     } catch {
                         onError(error)
                     }
@@ -161,9 +155,7 @@ extension DatabaseSnapshot {
             }
         case .unsafe:
             do {
-                if let value = try unsafeReentrantRead(observation.fetchFirst) {
-                    onChange(value)
-                }
+                try onChange(serializedDatabase.reentrantSync(observation.fetchValue))
             } catch {
                 onError(error)
             }